+++ apps/examples/src/examples/Playground.tsx
@@ -3,16 +3,15 @@ import { useMemo } from "react"
 import {
   bool,
   compileShader,
+  cos,
   float,
-  GLSLType,
-  Value,
-  variable,
+  join,
+  multiply,
+  sin,
+  SplitVector3,
   Variable,
-  vec2,
-  vec3,
-  vec4
+  vec3
 } from "shadenfreude"
-import { type } from "shadenfreude/src/glslType"
 import { Color, MeshStandardMaterial } from "three"
 import CustomShaderMaterial from "three-custom-shader-material"
 
@@ -22,88 +21,6 @@ const CSMMaster = (diffuseColor: Variable<"vec3">) =>
     inputs: { diffuseColor },
     fragmentBody: `csm_DiffuseColor = vec4(diffuseColor, 1.0);`
   })
-
-const Addition = <T extends GLSLType>(a: Value<T>, b: Value<T>) =>
-  variable(type(a), "a + b", {
-    title: `Addition (${type(a)} + ${type(b)}`,
-    inputs: { a, b }
-  })
-
-const Subtraction = <T extends GLSLType>(a: Value<T>, b: Value<T>) =>
-  variable(type(a), "a - b", {
-    title: `Subtraction (${type(a)} - ${type(b)}`,
-    inputs: { a, b }
-  })
-
-const Division = <T extends GLSLType>(a: Value<T>, b: Value<T>) =>
-  variable(type(a), "a / b", {
-    title: `Division (${type(a)} / ${type(b)}`,
-    inputs: { a, b }
-  })
-
-const Multiplication = <T extends GLSLType>(a: Value<T>, b: Value<T>) =>
-  variable(type(a), "a * b", {
-    title: `Multiplication (${type(a)} * ${type(b)}`,
-    inputs: { a, b }
-  })
-
-const add = Addition
-const subtract = Subtraction
-const divide = Division
-const multiply = Multiplication
-
-const sin = (x: Value<"float">) => float("sin(x)", { inputs: { x } })
-const cos = (x: Value<"float">) => float("cos(x)", { inputs: { x } })
-
-const SplitVector3 = (vector: Value<"vec3">) => ({
-  x: float("vector.x", { inputs: { vector } }),
-  y: float("vector.y", { inputs: { vector } }),
-  z: float("vector.z", { inputs: { vector } })
-})
-
-const JoinVector3 = (x: Value<"float">, y: Value<"float">, z: Value<"float">) =>
-  vec3("vec3(x, y, z)", {
-    inputs: { x, y, z }
-  })
-
-type Vector2Components = [Value<"float">, Value<"float">]
-type Vector3Components = [Value<"float">, Value<"float">, Value<"float">]
-type Vector4Components = [
-  Value<"float">,
-  Value<"float">,
-  Value<"float">,
-  Value<"float">
-]
-
-type JoinReturnType<Args> = Args extends Vector4Components
-  ? Variable<"vec4">
-  : Args extends Vector3Components
-  ? Variable<"vec3">
-  : Args extends Vector2Components
-  ? Variable<"vec2">
-  : never
-
-const join = <
-  Args extends Vector2Components | Vector3Components | Vector4Components
->(
-  ...args: Args
-) => {
-  const [x, y, z, w] = args
-
-  if (w !== undefined) {
-    return vec4("vec4(x, y, z, w)", {
-      inputs: { x, y, z, w }
-    }) as JoinReturnType<Args>
-  }
-
-  if (z !== undefined) {
-    return vec3("vec3(x, y, z)", { inputs: { x, y, z } }) as JoinReturnType<
-      Args
-    >
-  }
-
-  return vec2("vec2(x, y)", { inputs: { x, y } }) as JoinReturnType<Args>
-}
 
 const Time = () =>
   float("u_time", {
@@ -120,10 +37,7 @@ export default function Playground() {
     const v3 = join(1, 1, 1)
     const v4 = join(1, 1, 1, 1)
 
-    const { x } = SplitVector3(baseColor)
-    const time = Time()
-    const color = join(sin(time), cos(time), sin(multiply(time, 2)))
-    const root = CSMMaster(color)
+    const root = CSMMaster(baseColor)
 
     return compileShader(root)
   }, [])
