+++ packages/shadenfreude/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./lib/concatenator3000"
+export * from "./nodes"
 export * from "./variables"
 export { compileShader } from "./compilers"
