+++ packages/shadenfreude/src/nodes/index.ts
@@ -1,0 +1,2 @@
+export * from "./math"
+export * from "./vectors"
